+++ .github/workflows/claude-review-einops.yml
@@ -1,31 +1,20 @@
 name: "Claude Review: Einops"
 on:
-  pull_request:
-    types: [opened, reopened]
-    paths: ["**/*.py"]
-  workflow_dispatch:
+  workflow_call:
     inputs:
       pr_number:
-        description: "Pull Request number to review"
         required: true
         type: string
-
-permissions:
-  contents: read
-  pull-requests: write
-  id-token: write
 
 jobs:
   review:
     uses: ./.github/workflows/claude-review-base.yml
-    secrets:
-      anthropic_api_key: ${{ secrets.ANTHROPIC_API_KEY }}
     with:
       review_name: "Einops Suggestions"
       review_type: "einops"
       file_pattern: "\\.py$"
       tools: "Edit,Replace,Bash(git diff HEAD~1)"
-      pr_number: ${{ inputs.pr_number || github.event.pull_request.number || '' }}
+      pr_number: ${{ inputs.pr_number }}
       prompt: |
         Review the Python diff for PyTorch tensor operations that could be significantly improved
         with einops.rearrange.
